Update patch set 1
Patch Set 1:

Yes this is the new internal cm recovery (aka simplerecovery).

You can support this or CWM or both.  CWM is still the default recovery for now.  Not sure if/when that will change.

I assume all the pending CWM commits will eventually be submitted.  I've got a backlog of things to look at so it may be a bit of time before they all get done.

Patch-set: 1