Update patch set 3
Patch Set 3:

I dislike the assumption that primary sdcard *must* be named sdcard0.  I don't see anything in the storage spec that requires this, only examples that show it.  And, as you pointed out, goldfish is a counterexample.

How about we do the reverse -- assume that sdcard[1-9] is not primary storage, and anything else that begins with sdcard is primary storage.  Does that sound reasonable?

Patch-set: 3