Update patch set 1
Patch Set 1:

Ah, yeah, that makes sense. It does seem a little non-intuitive the way the fseeks are ordered though. While it works, it seems like you should first seek to the region where the bootloader message is and then seek to the specific member of the bootloader message.

Patch-set: 1