Update patch set 1
Patch Set 1:

I would leave the DUAL_DSI ifdef out and rely on sysfs. That macro was something I added so that I didn't need to shit up the display driver with ifdefs. The issue is that new kernels have support for partial updates on multiple controllers, and the kernel interface is different. This is a specific feature that was backported from 3.10 and that macro is the marker I added for it. You can still have a dual controller setup without it.

Patch-set: 1