Update patch set 1
Patch Set 1:

Most or all of the postrecoveryboot stuff is just resetting the boot-to-recovery flag.

That flag should be in the BCB (boot control block) on the misc partition on QCOM devices.  So if you add the misc partition to the fstab, it should work without the postrecoveryboot script.  Note that even within QCOM devices there are oddballs.  For example m7 requires a special offset, which we do support via a board flag.

Non-QCOM devices have different mechanisms.  For example, I have an encore that stores data in a vfat partition called boot.  There are patches up on gerrit to handle this, but I'm not sure if/when they will be merged.


Regardless, yes, 512mb devices that can run CM11 are the exception.  So I'd like to default to reasonable values for 1gb+ devices and let 512mb device maintainers override the defaults.  If there are enough of these devices, perhaps Phyto's suggestion of a board flag would be a good solution.

Patch-set: 1